10/02 finalizando as tarefas do git **lista de tarefas não aparece no database

+++ 00_APP_TO_DO_LIST/APP_TO_DO_LIST/Business/IBaseBusiness.cs
@@ -8,6 +8,6 @@ public interface IBaseBusiness<T>
     T Create(T item);
     T Update(T item);
     List<T> FindAll();
-    public T FindById(long id);
-    void Delete(long id);
+    public T FindById(int id);
+    void Delete(int id);
 }
